backend/requirements.txt
fastapi
uvicorn[standard]
websockets
sqlalchemy
alembic
psycopg2-binary   # si tu utilises PostgreSQL   
python-dotenv
